Name, colspan, display, structure
Extra tests for colspan and display (none),
Renamed many tests and put the information in alphabetical order for better maintenance.
Moved the asserts for function cloning to a separate tests so it's easier to test new code. Before cloning was tested at the start of the tests for a specific function. If the function was not a clone the test would fail. Now the edited function can be tested separately and while debugging the the clone-test can be ignored until the code is copied to the main file (I should create a better way to do this).